fix: address review comments for Hytte-4dc7h
