+++ web/src/hooks/useUnloadWarning.ts
@@ -1,0 +1,13 @@
+import { useEffect } from 'react'
+
+export function useUnloadWarning(active: boolean) {
+  useEffect(() => {
+    if (!active) return
+    function handleBeforeUnload(e: BeforeUnloadEvent) {
+      e.preventDefault()
+      e.returnValue = ''
+    }
+    window.addEventListener('beforeunload', handleBeforeUnload)
+    return () => window.removeEventListener('beforeunload', handleBeforeUnload)
+  }, [active])
+}
